pyproject.toml
[tool.poetry]
name = "comprag"
version = "0.1.0"
description = ""
authors = ["Osztopáni Kristóf <kristofosztopani@gmail.com>"]
license = "All Rights Reserved"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
python-dotenv = "^1.0.0"
langchain-chroma = "^0.1.1"
langchain-core = "^0.2.5"
langchain = "^0.2.3"
langchain-community = "^0.2.4"
langchain-text-splitters = "^0.2.1"
langchain-openai = "^0.1.8"
asgiref = "^3.8.1"
asyncio = "^3.4.3"
langchain-mistralai = "^0.1.8"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^6.1.0"
mypy = "^1.8.0"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
format = "commands:fmt"
lint = "commands:lint"
type-check = "commands:type_check"
check = "commands:check"

# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 0
line_length = 200

# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
[tool.black]
line-length = 200
target-version = ['py310']

[tool.flake8]
max-line-length = 200
exclude = [
    ".git",
    "__pycache__"
]
extend-ignore = []

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
show_error_codes = true
python_version = 3.10
exclude = [
    ".git",
    "__pycache__"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
